refactor: enforce index.ts as barrel-only files
Move implementation code from index.ts into dedicated files and
convert index.ts to pure re-export barrels:

- tools/api-goldens → extract-api-goldens.ts
- dashboards-ng module-federation/native-federation → register-loader.ts
- schematics migrations/data → element-migration-data.ts
- schematics migrations/ngx-translate → missing-translate-migration.ts
- schematics ng-update → migrate-to-v49.ts
- schematics ng-add → ng-add-rule.ts

Add ESLint no-restricted-syntax rule scoped to **/index.ts to prevent
future code declarations in barrel files.

Closes #1952
