+++ projects/element-ng/eslint.config.js
@@ -1,4 +1,4 @@
-import { tsConfig, templateConfig } from '../../eslint.config.js';
+import { tsConfig, templateConfig, indexBarrelConfig } from '../../eslint.config.js';
 import defaultValuePlugin from '@siemens/eslint-plugin-defaultvalue';
 import { defineConfig, globalIgnores } from 'eslint/config';
 export default defineConfig(
@@ -43,5 +43,6 @@ export default defineConfig(
       '@typescript-eslint/no-deprecated': ['off']
     }
   },
-  ...templateConfig
+  ...templateConfig,
+  ...indexBarrelConfig
 );
