+++ projects/element-ng/schematics/ng-add/ng-add-rule.ts
@@ -1,0 +1,19 @@
+/**
+ * Copyright (c) Siemens 2016 - 2026
+ * SPDX-License-Identifier: MIT
+ */
+import { chain, Rule, schematic, SchematicContext, Tree } from '@angular-devkit/schematics';
+import { getPackageJsonDependency } from '@schematics/angular/utility/dependencies';
+
+export const ngAdd = (options: { path: string }): Rule => {
+  return (tree: Tree, context: SchematicContext) => {
+    context.logger.info('🔧 Adding @siemens/element-ng to your project...');
+
+    const hasSimplElementNgDependency = getPackageJsonDependency(tree, '@simpl/element-ng');
+
+    if (hasSimplElementNgDependency) {
+      const chainedRules = chain([schematic('simpl-siemens-migration', options)]);
+      return chainedRules(tree, context);
+    }
+  };
+};
